WIP: ADR-053 Stage 4 — hoist service assembly into the daemon library
Moves SharedServices, DatabaseServices, build_shared_services,
build_database_services (and the seed_agent_nodes / resolve_model_path /
load_shared_embedding_model_bg / wire_database_embeddings_bg helpers) out of the
`nodespaced` binary and into a new `services::assembly` library module, so the
DatabaseManager can build and cache per-database service sets in a later
increment. Pure relocation — no behavior change.

## Completed in This Session
- [x] New packages/daemon/src/services/assembly.rs with the structs (fields now
      `pub`) + build functions, re-exported from services/mod.rs and lib.rs.
- [x] main.rs calls `nodespace_daemon::build_shared_services` /
      `build_database_services` and drops the now-unused imports; `drain_gpu`
      and the serve loops are otherwise unchanged.
- [x] cargo fmt/clippy(-D warnings) clean; daemon tests 74/74 (lib 45 + main 1 +
      grpc_round_trip 16 + agent_session 9 + import 3) — the reassembled services
      still pass the end-to-end suites.

## Remaining Work (Stage 4)
- [ ] Give DatabaseManager a shared build-context (pty_manager + model watch +
      has_model; settings stays router-level) and change build_database_services
      to take it; wire get_or_open to build+cache Arc<DatabaseServices> in the
      open-map (replace the () placeholder); ensure_default_registered on boot.
- [ ] Tonic interceptor: read x-ns-database-id, resolve_database_id, inject the
      resolved Arc<DatabaseServices> as a request extension; header-less ->
      default. Handlers read their per-database services from the extension.

## Current State
- Files: new services/assembly.rs; edits to services/mod.rs, lib.rs, main.rs.
- Build/tests: cargo build/fmt/clippy -D warnings clean; daemon 74/74.
- Known issues: none. DatabaseManager.open still holds (); get_or_open still
  returns "not yet wired" — next increment.

## Context for Next Session
The assembly now lives in the lib. Next: split SharedServices into a
router-owned `settings` and a cloneable build-context, hand the context to
DatabaseManager so get_or_open can call build_database_services and cache the
result, then add the x-ns-database-id interceptor.
